move jobtemplate classes to own file
